sdk/neuroflow/templates/agent/advanced/requirements.txt
# Advanced Template Dependencies
neuroflow-sdk>=0.4.1

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Configuration
pyyaml>=6.0

# HTTP Client
aiohttp>=3.8.0

# MCP (optional)
# @modelcontextprotocol/sdk

# Advanced Features
redis>=4.0.0  # 缓存
prometheus-client>=0.16.0  # 指标

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
